workspace-agent/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
apscheduler==3.10.4
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2